requirements.txt
numpy>=1.24
matplotlib>=3.7
pyyaml>=6.0
networkx>=3.0
